requirements.txt
altgraph==0.17.4
et_xmlfile==2.0.0
importlib_metadata==8.5.0
macholib==1.16.3
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
ttkbootstrap==1.10.1
zipp==3.21.0
